requirements.txt
requests>=2.32.3
beautifulsoup4>=4.12.3
lxml>=5.2.2
tldextract>=5.1.2
python-whois>=0.9.4
dnspython>=2.6.1
python-dateutil>=2.9.0
python-dotenv>=1.0.1
fastapi>=0.112.0
uvicorn[standard]>=0.30.3
python-multipart>=0.0.9
pydantic>=2.7.4
